+++ app/Console/Kernel.php
@@ -24,9 +24,22 @@ class Kernel extends ConsoleKernel
      */
     protected function schedule(Schedule $schedule)
     {
-        $schedule->command('telescope:prune --hours=72')->daily();
-        $schedule->command('backup:clean')->daily()->at('01:00');
-        $schedule->command('backup:run')->daily()->at('02:00');
+        $schedule->command('soloc:moderator-queue-notification')
+                ->weekly()
+                ->mondays()
+                ->thursdays()
+                ->at('8:00');
+
+        $schedule->command('telescope:prune --hours=72')
+                ->daily();
+
+        $schedule->command('backup:clean')
+                ->daily()
+                ->at('01:00');
+
+        $schedule->command('backup:run')
+                ->daily()
+                ->at('02:00');
     }
 
     /**
@@ -39,5 +52,15 @@ class Kernel extends ConsoleKernel
         $this->load(__DIR__.'/Commands');
 
         require base_path('routes/console.php');
+    }
+
+    /**
+     * Get the timezone that should be used by default for scheduled events.
+     *
+     * @return \DateTimeZone|string|null
+     */
+    protected function scheduleTimezone()
+    {
+        return 'America/Montreal';
     }
 }